Sample image save error debug

+++ LungCancer/02_testing/dcgan_imagenet.py
@@ -26,8 +26,8 @@ def merge(images, size):
     img = np.zeros((h * size[0], w * size[1]))
 
     for idx, image in enumerate(images):
-        i = idx % size[1]
-        j = idx / size[1]
+        i = int(idx % size[1])
+        j = int(idx / size[1])
         print ("i: {0}, j: {1}".format(i, j))
         img[j * h:j * h + h, i * w:i * w + w] = image
     return img
